Use dep-info to decide if output is cacheable
Cargo ask's rustc to emit a dep-info file, so we can piggy-back on that.
After calling rustc, we check if any of the listed files have been
modified since the build started. If so, we conclude that the result
is not cachable because we don't know exactly what contributed to it.

Now that I know what's cacheable, the next step is to store that info
somewhere and use it from other Cargo builds.