Adopt cached build artifacts if we find them
This is the first end-to-end demonstration of this
giant pile of hacks actually working... and it does!
Or at least it appears to. :)

I got impatient in this bit and did a bunch of
terrible copy+pasta, so my next move will be to
clean all of that up, and then try to improve visibility.
(Might be able to communicate messages back to Cargo;
not sure how to do that.)