Get package names correctly
And add long-form rustc command line arguments that were missing from
the online reference.

Also add some extra logging.

I can get a rather large project to build with this now, including adopting
some dependencies that were built for another project, but it doesn't seem
to be having a huge impact on build time. I realised that one of my biggest
dependencies is stuck on an older version than the "source" project,
so I'm going to try bringing that up-to-date and see if I get more cache hits
for the big stuff.