Add start of integration tests
I'm going to go a different direction where I only cache
build output for crates that come from immutable sources.
This was inspired by discussions about plans for first-class
caching support in Cargo itself. The idea is that these
are much easier to cache because the fingerprint is all
you need.

So next up I'm going to rip out all the stuff around mtimes
and so on. I could always add it back later if the simpler
version of this turns out to be useful!