Don't cache build script output
Instead skip building and running build scripts entirely
if we know we can pull the final crate output, and just
emit the build script's stdout from cache.

Note that this temporarily makes things slightly worse in
some ways, in that Cargo will attempt to rebuild the build
scripts over and over because it will _never_ find the
files that have been emitted in "rerun-if-changed" directives.
To deal with this, I'm intending to filter them out when
we print stdout we got from the cache. But it's not totally
now so I'm going to get this rework onto the main branch
sooner rather than later.