Second attempt at deferred build script execution
Build scripts are replaced with a copy of Hope itself,
which detects when it is invoked as a build script
and only eagerly invokes the real build script if
absolutely necessary. (This is if we can't find stdout
for a matching build script invocation in the cache.)

Then when Hope is invoked for building the main crate,
iff we actually need to build the main crate then we
will invoke the _real_ build script with the environment
it was originally called with, and rewind timestamps
of any files created to look like they were created
when Cargo first thought it was running the build script.

There are some other details I haven't documented in
this commit message, but in my (limited) testing it
actually seems to work pretty well now!